[FIX] New Organization on add-bank-account feature
